Add support for training sampler
This is needed to use a WeightedRandomSampler to support oversampling
of rare class, for example. The sampler can't simply be passed in as an
kwarg to databunch() because we only want to pass it to the train_dl,
and we need to turn off the shuffle for the train_dl because that option is
not compatible with setting a sampler.